requirements.txt
aiofiles==24.1.0
anyio==4.10.0
async-property==0.2.2
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
dash==3.0.4
dash-bootstrap-components==2.0.3
dash-iconify==0.1.2
dash_mantine_components==2.0.0
deprecation==2.1.0
Flask==3.0.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.2
narwhals==1.41.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
plotly==6.1.2
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keycloak==5.8.1
pytz==2025.2
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.22.0
geopandas==1.1.1